[#1300] Don't stream aggregate's CSV responses
Although it would be very important to stream the response, as CSVs can get big,
we can't do so with our current caching infrastructure that relies on pickling
the responses. Streaming responses can't be pickled.

I've done this change in a way that's trivial to come back to streaming,
whenever our caching infrastructure supports it.

openspending/openspending#1300